+++ CHANGELOG.md
@@ -99,6 +99,8 @@ All notable changes to this project will be documented in this file.
 - split instagram feed actions file into create and update action files
 - feed form hook to use tanstack query mutation
 - feed list to fetch from supabase
+- feed form to fetch selected images for feed from supabase
+- feed form to fetch media from supabase
 
 ###### Media